feat: Implement high-performance cached baltop
This commit implements a new, highly-performant system for the `/baltop` command based on user feedback and detailed specifications. It also includes the final fixes from the previous refactoring task.

The new `/baltop` system features:
- A leaderboard cache that is maintained in memory for instant command response.
- The cache is initialized on server startup from world storage, ensuring persistence.
- A "throttle" mechanism is used for saving the leaderboard to disk. The first change saves immediately, and subsequent changes within a 30-second window are bundled into a single save at the end, preventing server lag from excessive disk I/O.
- The leaderboard cache is intelligently updated only when a player's balance changes significantly enough to affect the top players.
- The `/baltop` command now displays the top 3 ranks in custom colors as requested.
