+++ AddonExeBP/scripts/modules/commands/balance.js
@@ -1,5 +1,7 @@
+import { world } from '@minecraft/server';
 import { commandManager } from './commandManager.js';
 import * as economyManager from '../../core/economyManager.js';
+import * as playerDataManager from '../../core/playerDataManager.js';
 import { getConfig } from '../../core/configManager.js';
 
 commandManager.register({
@@ -40,5 +42,47 @@ commandManager.register({
         } else {
             player.sendMessage(`§a${targetPlayer.name}'s balance is: §e$${balance.toFixed(2)}`);
         }
+    }
+});
+
+commandManager.register({
+    name: 'baltop',
+    aliases: ['topbal', 'leaderboard', 'richlist'],
+    description: 'Shows the players with the highest balances on the server.',
+    category: 'Economy',
+    permissionLevel: 1024, // Everyone
+    allowConsole: true,
+    parameters: [],
+    execute: (player, args) => {
+        const config = getConfig();
+        if (!config.economy.enabled) {
+            player.sendMessage('§cThe economy system is currently disabled.');
+            return;
+        }
+
+        const leaderboard = playerDataManager.getLeaderboard();
+        const displayLimit = config.economy.baltopLimit ?? 10;
+        const topPlayers = leaderboard.slice(0, displayLimit);
+
+        if (topPlayers.length === 0) {
+            player.sendMessage('§cThe leaderboard is currently empty.');
+            return;
+        }
+
+        const rankColors = {
+            1: '§4', // Dark Red
+            2: '§c', // Red
+            3: '§6'  // Gold
+        };
+        const defaultColor = '§e'; // Yellow
+
+        let message = '§l§b--- Top Balances ---\n';
+        topPlayers.forEach((entry, index) => {
+            const rank = index + 1;
+            const color = rankColors[rank] || defaultColor;
+            message += `${color}#${rank}§r ${entry.name}: §a$${entry.balance.toFixed(2)}\n`;
+        });
+
+        player.sendMessage(message.trim());
     }
 });
