--- Dev/tasks/completed.md
@@ -1,5 +1,0 @@
-# Completed Development Tasks
-
-This document is an archive of completed tasks.
-
----
